Updates output_viewer
I had an issue where the diag viewer would fail to recognize the package correctly, and fail to upload even though everything was in the right place. Turns out it chdirs to the target directory (/something/something/amwg/) and then tries to upload the files 'amwg/FILE', which doesnt work because its already in the amwg directory. By moving up a directory, it works.